Add test, refactor to make it more testable
